svn path=/branches/arty-newcc/; revision=53423

+++ drivers/filesystems/cdfs/finfo.c
@@ -153,7 +153,7 @@ CdfsGetNameInformation(PFILE_OBJECT FileObject,
     ASSERT(Fcb != NULL);
 
     /* If buffer can't hold at least the file name length, bail out */
-    if (*BufferLength < FIELD_OFFSET(FILE_NAME_INFORMATION, FileName[0]))
+    if (*BufferLength < (ULONG)FIELD_OFFSET(FILE_NAME_INFORMATION, FileName[0]))
         return STATUS_BUFFER_OVERFLOW;
 
     /* Calculate file name length in bytes */
